+++ core/src/parsing/formula.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::substance::{Properties, Property, Substance};
+use crate::runtime::{Properties, Property, Substance};
 use crate::types::{Dimension, Number, Numeric};
 use std::collections::BTreeMap;
 use std::iter::Peekable;
