+++ core/src/runtime/mod.rs
@@ -3,9 +3,11 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 mod eval;
+mod substance;
 mod value;
 
 pub(crate) use eval::{eval_expr, eval_query};
 pub(crate) use value::Show;
 
+pub use substance::{Properties, Property, Substance, SubstanceGetError};
 pub use value::Value;
