+++ core/src/runtime/substance.rs
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use super::Show;
 use crate::loader::Context;
 use crate::output::{PropertyReply, SubstanceReply};
-use crate::runtime::Show;
 use crate::types::{Digits, Dimension, Number, Numeric};
 use std::collections::BTreeMap;
 use std::iter::once;
